What is GitOps: mention tools required for convergence of state
N.B.: tools are mentioned in the "Key benefits of GitOps" section
of https://www.weave.works/technologies/gitops/#key-benefits:

> "it uses tools to compare the actual production state of your
> whole application with what’s under source control and then it
> tells you when your cluster doesn’t match the real world."

but this didn't make it to the definition.
I've added it as Git only doesn't make things GitOps-ey. If there
isn't any automated process in place for convergence, then it
could be ManualOps, CIOps, or any other way to do CD.